Doct: Tenemos la cookie(guardamos el email), ya tenemos bien la contraseña en el log-in; el shopping cart esta malo(Json basio)

+++ src/server/popos/Product.php
@@ -31,7 +31,7 @@ class Product{
     }
 
     // Setter para price
-    public function setPrice(int $price): void {
+    public function setPrice(float $price): void {
         $this->price = $price;
     }
